Extract ZoneMap into its own module
